+++ spec/support/cookbooks/private_internet_access_test/recipes/default.rb
@@ -1,0 +1,5 @@
+# Encoding: UTF-8
+
+private_internet_access 'default' do
+  source node['private_internet_access']['app']['source']
+end